move / re-move some stuff
